Move methods to Importer class. Define 'Block' in linescanner.py
