+++ leo/plugins/importers/c.py
@@ -4,12 +4,11 @@
 from __future__ import annotations
 import re
 from typing import List, Tuple, TYPE_CHECKING
-from leo.plugins.importers.linescanner import Importer, ImporterError
+from leo.plugins.importers.linescanner import Block, Importer, ImporterError
 from leo.core import leoGlobals as g
 if TYPE_CHECKING:
     from leo.core.leoCommands import Commands as Cmdr
     from leo.core.leoNodes import Position
-    Block = Tuple[str, str, int, int, int]  # (kind, name, start, start_body, end)
 
 #@+others
 #@+node:ekr.20140723122936.17928: ** class C_Importer
@@ -32,23 +31,6 @@ class C_Importer(Importer):
             # 'break', 'continue',
         ])
         self.c_keywords_pattern = re.compile(self.c_keywords)
-    #@+node:ekr.20230513080610.1: *3* c_i.common_indent (move to base class)
-    def common_indent(self, blocks: List[Block]) -> int:
-        """
-        Return the length of the common leading indentation of
-        all non-blank lines in all blocks.
-        """
-        if not blocks:
-            return 0
-        lws_list: List[int] = []
-        for block in blocks:
-            kind, name, start, start_body, end = block
-            lines = self.lines[start:end]
-            for line in lines:
-                stripped_line = line.lstrip()
-                if stripped_line:  # Skip empty lines
-                    lws_list.append(len(line[: -len(stripped_line)]))
-        return min(lws_list) if lws_list else 0
     #@+node:ekr.20220728055719.1: *3* c_i.find_blocks
     #@+<< define block_patterns >>
     #@+node:ekr.20230511083510.1: *4* << define block_patterns >>
@@ -151,7 +133,7 @@ class C_Importer(Importer):
             # Start with the head: lines[start : start_start_body].
             result_list = lines[start:start_body]
             # Add indented @others.
-            common_lws_s = ' ' * self.common_indent(blocks)
+            common_lws_s = ' ' * self.compute_common_lws(blocks)
             result_list.extend([f"{common_lws_s}@others\n"])
 
             # Recursively generate the inner nodes/blocks.
@@ -200,20 +182,6 @@ class C_Importer(Importer):
 
         # Add trailing lines.
         parent.b += f"@language {self.name}\n@tabwidth {self.tab_width}\n"
-    #@+node:ekr.20230513085654.1: *3* c_i.remove_common_lws (move to base class)
-    def remove_common_lws(self, lws: str, p: Position) -> None:
-        """Remove the given leading whitespace from all lines of p.b."""
-        if len(lws) == 0:
-            return
-        assert lws.isspace(), repr(lws)
-        n = len(lws)
-        lines = g.splitLines(p.b)
-        result: List[str] = []
-        for line in lines:
-            stripped_line = line.strip()
-            assert not stripped_line or line.startswith(lws), repr(line)
-            result.append(line[n:] if stripped_line else line)
-        p.b = ''.join(result)
     #@-others
 #@-others
